# LibRegex: Various performance improvements