Before:

Finished 0 tests and 31 benchmarks in 10451ms (0ms tests, 10435ms benchmarks, 16ms other).

After:

Finished 0 tests and 31 benchmarks in 7084ms (0ms tests, 7069ms benchmarks, 15ms other).

**LibRegex: Use a plain array to store opcodes**

Using a hash map is unnecessary because the number of opcodes and their IDs never change.

**LibRegex: Remove return value for setters**

**LibRegex: Make get_opcode() return a reference**

Previously this would return a pointer which could be null if the requested opcode was invalid. This should never be the case though so let's `VERIFY()` that instead.

**LibRegex: Avoid making unnecessary string copies**

**LibRegex: Avoid prepending items to vectors**

**LibRegex: Avoid initialization checks in get_opcode_by_id()**

**LibRegex: Use a plain pointer for OpCode::m_state**

**LibRegex: Remove unused code**